fakevm-agent-ryu: add check tunnel device type
In the multi-node-emulation mode, restricts type of target device which
is given a tunnel IP address at initialize or is removed at clean.

+++ quantum/fakevm/fakevm_agent_ryu.py
@@ -16,6 +16,8 @@
 #    under the License.
 # @author: Isaku Yamahata
 
+import os.path
+
 from oslo.config import cfg
 
 from quantum.agent.linux import ip_lib
@@ -83,23 +85,27 @@ class QuantumFakeVMAgentRyu(
         self._cleanup_tunnel()
         dev_name = self.conf.OVS.tunnel_interface
         ip_wrapper = ip_lib.IPWrapper(self.root_helper)
+        device = None
         if not ip_lib.device_exists(dev_name, self.root_helper):
             # ip link add $tunnel_interface type dummy
             device = ip_wrapper.add_dummy(dev_name)
-        else:
+        elif os.path.exists('/sys/devices/virtual/net/%s' % devname):
             device = ip_wrapper.device(dev_name)
-        if self.conf.OVS.tunnel_ip:
+        if device and self.conf.OVS.tunnel_ip:
             # ip address add $tunnel_ip brd '+' scope global dev $dev_name
             device.addr.add(4, self.conf.OVS.tunnel_ip, '+')
             # ip link set $tunnel_interface up
             device.link.set_up()
+            self._execute(['ip', 'route', 'add', self.conf.OVS.tunnel_ip,
+                           'dev', dev_name])
 
         self._ensure_ovs_bridge(self.conf.OVS.integration_bridge)
 
     def _cleanup_tunnel(self):
         dev_name = self.conf.OVS.tunnel_interface
         if dev_name and self.conf.OVS.tunnel_ip:
-            if ip_lib.device_exists(dev_name, self.root_helper):
+            if (ip_lib.device_exists(dev_name, self.root_helper) and
+                os.path.exists('/sys/devices/virtual/net/%s' % dev_name)):
                 ip_wrapper = ip_lib.IPWrapper(self.root_helper)
                 ip_wrapper.device(dev_name).link.delete()
